break apart models so it's not such a massive blob of definitions
